chore(assets-retry): compile runtime code via Rsbuild plugin
